fs: introduce dmemfs module
dmemfs (Direct Memory filesystem) is device memory or reserved
memory based filesystem. This kind of memory is special as it
is not managed by kernel and it is not associated with 'struct page'.

The original purpose for dmemfs is to drop the usage of
'struct page' to save extra system memory in public cloud
enviornment.

This patch introduces the basic framework of dmemfs and only
mkdir and create regular file are supported.

Signed-off-by: Chen Zhuo <sagazchen@tencent.com>